Use long instead of int for chat ids.
This fix comes from: https://github.com/apimorphism/telegramium/pull/5
by @oybek.

+++ telegramium-core/src/main/scala/telegramium/bots/ChatId.scala
@@ -2,6 +2,6 @@ package telegramium.bots
 
 sealed trait ChatId {}
 
-final case class ChatIntId(id: Int) extends ChatId
+final case class ChatIntId(id: Long) extends ChatId
 
 final case class ChatStrId(id: String) extends ChatId
