+++ telegramium-core/src/main/scala/telegramium/bots/ResponseParameters.scala
@@ -9,7 +9,7 @@ final case class ResponseParameters(
                                       * smaller than 52 bits, so a signed 64 bit integer or
                                       * double-precision float type are safe for storing this
                                       * identifier.*/
-                                    migrateToChatId: Option[Int] = Option.empty,
+                                    migrateToChatId: Option[Long] = Option.empty,
                                     /** Optional. In case of exceeding flood control, the number of
                                       * seconds left to wait before the request can be repeated*/
                                     retryAfter: Option[Int] = Option.empty)
